requirements_annotate.txt
google-genai>=1.0.0
Pillow>=10.0.0
pydantic>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
sam2
huggingface_hub
